// back/src/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };

/** All built-in and custom scalars, mapped to their actual values */

export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type AddArticleResponse = {
  __typename?: 'AddArticleResponse';
  article?: Maybe<Article>;
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type AddCommentResponse = {
  __typename?: 'AddCommentResponse';
  code: Scalars['Int']['output'];
  comment?: Maybe<Comment>;
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type AddLikeResponse = {
  __typename?: 'AddLikeResponse';
  code: Scalars['Int']['output'];
  like?: Maybe<Like>;
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type Article = {
  __typename?: 'Article';
  author: User;
  content: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  likes?: Maybe<Array<Maybe<Like>>>;
  title: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['String']['output']>;
};

export type ArticleDetails = {
  __typename?: 'ArticleDetails';
  author: User;
  comments?: Maybe<Array<Maybe<Comment>>>;
  content: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  likes?: Maybe<Array<Maybe<Like>>>;
  title: Scalars['String']['output'];
};

export type Comment = {
  __typename?: 'Comment';
  author: User;
  content: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  id: Scalars['ID']['output'];
};

export type CreateUserResponse = {
  __typename?: 'CreateUserResponse';
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
  user?: Maybe<User>;
};

export type DeleteArticleResponse = {
  __typename?: 'DeleteArticleResponse';
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type DeleteCommentResponse = {
  __typename?: 'DeleteCommentResponse';
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type DeleteLikeResponse = {
  __typename?: 'DeleteLikeResponse';
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type GetArticleResponse = {
  __typename?: 'GetArticleResponse';
  article?: Maybe<ArticleDetails>;
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type GetArticlesResponse = {
  __typename?: 'GetArticlesResponse';
  articles?: Maybe<Array<Maybe<Article>>>;
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type GetAuthorsResponse = {
  __typename?: 'GetAuthorsResponse';
  authors?: Maybe<Array<Maybe<User>>>;
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type Like = {
  __typename?: 'Like';
  id: Scalars['ID']['output'];
  user: User;
};

export type LogInResponse = {
  __typename?: 'LogInResponse';
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
  token?: Maybe<Scalars['String']['output']>;
  user?: Maybe<User>;
};

export type Mutation = {
  __typename?: 'Mutation';
  addArticle?: Maybe<AddArticleResponse>;
  addComment?: Maybe<AddCommentResponse>;
  addLike?: Maybe<AddLikeResponse>;
  createUser?: Maybe<CreateUserResponse>;
  deleteArticle?: Maybe<DeleteArticleResponse>;
  deleteComment?: Maybe<DeleteCommentResponse>;
  deleteLike?: Maybe<DeleteLikeResponse>;
  logIn?: Maybe<LogInResponse>;
  updateArticle?: Maybe<UpdateArticleResponse>;
};
// ...
export type Query = {
  __typename?: 'Query';
  getArticle?: Maybe<GetArticleResponse>;
  getArticles?: Maybe<GetArticlesResponse>;
  getAuthors?: Maybe<GetAuthorsResponse>;
  me?: Maybe<GetUserResponse>;
};


export type QueryGetArticleArgs = {
  id: Scalars['ID']['input'];
};


export type QueryGetArticlesArgs = {
  authorId?: InputMaybe<Scalars['String']['input']>;
  orderByLikesAsc?: InputMaybe<Scalars['Boolean']['input']>;
  orderByLikesDesc?: InputMaybe<Scalars['Boolean']['input']>;
};

export type UpdateArticleProps = {
  content?: InputMaybe<Scalars['String']['input']>;
  title?: InputMaybe<Scalars['String']['input']>;
};

export type UpdateArticleResponse = {
  __typename?: 'UpdateArticleResponse';
  article?: Maybe<Article>;
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type User = {
  __typename?: 'User';
  id: Scalars['ID']['output'];
  username: Scalars['String']['output'];
};

export type GetUserResponse = {
  __typename?: 'getUserResponse';
  code: Scalars['Int']['output'];
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
  user?: Maybe<User>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = {
  AddArticleResponse: ResolverTypeWrapper<Omit<AddArticleResponse, 'article'> & { article?: Maybe<ResolversTypes['Article']> }>;
  AddCommentResponse: ResolverTypeWrapper<Omit<AddCommentResponse, 'comment'> & { comment?: Maybe<ResolversTypes['Comment']> }>;
  AddLikeResponse: ResolverTypeWrapper<Omit<AddLikeResponse, 'like'> & { like?: Maybe<ResolversTypes['Like']> }>;
  Article: ResolverTypeWrapper<Article>;
  ArticleDetails: ResolverTypeWrapper<Omit<ArticleDetails, 'comments' | 'likes'> & { comments?: Maybe<Array<Maybe<ResolversTypes['Comment']>>>, likes?: Maybe<Array<Maybe<ResolversTypes['Like']>>> }>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  Comment: ResolverTypeWrapper<Comment>;
  CreateUserResponse: ResolverTypeWrapper<CreateUserResponse>;
  DeleteArticleResponse: ResolverTypeWrapper<DeleteArticleResponse>;
  DeleteCommentResponse: ResolverTypeWrapper<DeleteCommentResponse>;
  DeleteLikeResponse: ResolverTypeWrapper<DeleteLikeResponse>;
  GetArticleResponse: ResolverTypeWrapper<Omit<GetArticleResponse, 'article'> & { article?: Maybe<ResolversTypes['ArticleDetails']> }>;
  GetArticlesResponse: ResolverTypeWrapper<Omit<GetArticlesResponse, 'articles'> & { articles?: Maybe<Array<Maybe<ResolversTypes['Article']>>> }>;
  GetAuthorsResponse: ResolverTypeWrapper<GetAuthorsResponse>;
  ID: ResolverTypeWrapper<Scalars['ID']['output']>;
  Int: ResolverTypeWrapper<Scalars['Int']['output']>;
  Like: ResolverTypeWrapper<Like>;
  LogInResponse: ResolverTypeWrapper<LogInResponse>;
  Mutation: ResolverTypeWrapper<{}>;
  Query: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
  UpdateArticleProps: UpdateArticleProps;
  UpdateArticleResponse: ResolverTypeWrapper<Omit<UpdateArticleResponse, 'article'> & { article?: Maybe<ResolversTypes['Article']> }>;
  User: ResolverTypeWrapper<User>;
  getUserResponse: ResolverTypeWrapper<GetUserResponse>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  AddArticleResponse: Omit<AddArticleResponse, 'article'> & { article?: Maybe<ResolversParentTypes['Article']> };
  AddCommentResponse: Omit<AddCommentResponse, 'comment'> & { comment?: Maybe<ResolversParentTypes['Comment']> };
  AddLikeResponse: Omit<AddLikeResponse, 'like'> & { like?: Maybe<ResolversParentTypes['Like']> };
  Article: Article;
  ArticleDetails: Omit<ArticleDetails, 'comments' | 'likes'> & { comments?: Maybe<Array<Maybe<ResolversParentTypes['Comment']>>>, likes?: Maybe<Array<Maybe<ResolversParentTypes['Like']>>> };
  Boolean: Scalars['Boolean']['output'];
  Comment: Comment;
  CreateUserResponse: CreateUserResponse;
  DeleteArticleResponse: DeleteArticleResponse;
  DeleteCommentResponse: DeleteCommentResponse;
  DeleteLikeResponse: DeleteLikeResponse;
  GetArticleResponse: Omit<GetArticleResponse, 'article'> & { article?: Maybe<ResolversParentTypes['ArticleDetails']> };
  GetArticlesResponse: Omit<GetArticlesResponse, 'articles'> & { articles?: Maybe<Array<Maybe<ResolversParentTypes['Article']>>> };
  GetAuthorsResponse: GetAuthorsResponse;
  ID: Scalars['ID']['output'];
  Int: Scalars['Int']['output'];
  Like: Like;
  LogInResponse: LogInResponse;
  Mutation: {};
  Query: {};
  String: Scalars['String']['output'];
  UpdateArticleProps: UpdateArticleProps;
  UpdateArticleResponse: Omit<UpdateArticleResponse, 'article'> & { article?: Maybe<ResolversParentTypes['Article']> };
  User: User;
  getUserResponse: GetUserResponse;
};
// ...
export type QueryResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  getArticle?: Resolver<Maybe<ResolversTypes['GetArticleResponse']>, ParentType, ContextType, RequireFields<QueryGetArticleArgs, 'id'>>;
  getArticles?: Resolver<Maybe<ResolversTypes['GetArticlesResponse']>, ParentType, ContextType, Partial<QueryGetArticlesArgs>>;
  getAuthors?: Resolver<Maybe<ResolversTypes['GetAuthorsResponse']>, ParentType, ContextType>;
  me?: Resolver<Maybe<ResolversTypes['getUserResponse']>, ParentType, ContextType>;
};